+++ src/libslic3r/EdgeGrid.hpp
@@ -78,8 +78,8 @@ protected:
 #endif
 	bool cell_inside_or_crossing(int r, int c) const
 	{
-		if (r < 0 || r >= m_rows ||
-			c < 0 || c >= m_cols)
+		if (r < 0 || (size_t)r >= m_rows ||
+			c < 0 || (size_t)c >= m_cols)
 			// The cell is outside the domain. Hoping that the contours were correctly oriented, so
 			// there is a CCW outmost contour so the out of domain cells are outside.
 			return false;
